requirements.txt
requests
Flask==2.0.2
Werkzeug==2.0.3
openai==1.35.7
python-dotenv==0.19.0